WIP on util for converting runtime args to utf8
