+++ core/src/nucleus/ribosome/get.rs
@@ -1,6 +1,7 @@
 use agent::ActionResult;
 use nucleus::ribosome::{HcApiReturnCode, Runtime};
 use serde_json;
+use super::runtime_args_to_utf8;
 use std::sync::mpsc::channel;
 use wasmi::{RuntimeArgs, RuntimeValue, Trap};
 
@@ -10,29 +11,9 @@ struct GetArgs {
 }
 
 pub fn invoke_get(runtime: &mut Runtime, args: &RuntimeArgs) -> Result<Option<RuntimeValue>, Trap> {
-    // @TODO assert or return error?
-    // @see https://github.com/holochain/holochain-rust/issues/159
-    assert!(args.len() == 2);
-
-    // assert!(false);
-
-    // Read complex argument serialized in memory
-    // @TODO use our Malloced data instead
-    // @see https://github.com/holochain/holochain-rust/issues/65
-    let mem_offset: u32 = args.nth(0);
-    let mem_len: u32 = args.nth(1);
-    let bin_arg = runtime
-        .memory
-        .get(mem_offset, mem_len as usize)
-        // @TODO panic here?
-        // @see https://github.com/holochain/holochain-rust/issues/159
-        .expect("Successfully retrive the arguments");
-
-    // deserialize complex argument
-    // @TODO panic here?
-    // @see https://github.com/holochain/holochain-rust/issues/159
-    let arg = String::from_utf8(bin_arg).unwrap();
-    let res_entry: Result<GetArgs, _> = serde_json::from_str(&arg);
+    // deserialize args
+    let args_str = runtime_args_to_utf8(&runtime, &args);
+    let res_entry: Result<GetArgs, _> = serde_json::from_str(&args_str);
     // Exit on error
     if res_entry.is_err() {
         // Return Error code in i32 format
@@ -78,7 +59,7 @@ pub fn invoke_get(runtime: &mut Runtime, args: &RuntimeArgs) -> Result<Option<Ru
             // TODO #65 - use our Malloc instead
             runtime
                 .memory
-                .set(mem_offset, &params)
+                .set(0, &params)
                 .expect("memory should be writable");
 
             // Return success in i32 format
